compartmentalized main.dar
Moved ToDo list functionality to separate todo.dart file
